pyproject.toml
[project]
name = "mas-aws"
version = "0.1.0"
description = "Multi-Agent RAG System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "strands-agents>=0.1.0",
    "openai>=1.0.0",
    "boto3>=1.35.0",
    "psycopg2-binary>=2.9.0",
    "qdrant-client[fastembed]>=1.12.0",
    "python-telegram-bot>=21.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "watchtower>=3.0.0",
]
